correccion de eroor de float a double en modelo producto

+++ IDGS903_Api/Models/Producto.cs
@@ -9,11 +9,11 @@ namespace IDGS903_Api.Models
         public String? Nombre { get; set; }
         public String? Descripcion { get; set; }
         public int Stock { get; set; }
-        public float Altura { get; set; }
-        public float Ancho { get; set; }
-        public float Largo { get; set; }
+        public Double Altura { get; set; }
+        public Double Ancho { get; set; }
+        public Double Largo { get; set; }
         public String? Imagen { get; set; }
-        public float Total { get; set; }
+        public Double Total { get; set; }
         public int Estatus { get; set; }
     }
 }
